+++ CHANGELOG.md
@@ -17,10 +17,10 @@ DEPENDENCIES:
 * Bumped github.com/hashicorp/terraform-plugin-mux from 0.23.0 to 0.23.1 ([#486])
 
 [#432]: https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/432
+[#453]: https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/453
 [#482]: https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/482
 [#485]: https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/485
 [#486]: https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/486
-[#453]: https://github.com/cloudamqp/terraform-provider-cloudamqp/pull/453
 
 ## 1.44.4 (14 Apr, 2026)